Support multiple inheritance, optimize edge classes
Every Base-derived class has a '_data_class' property that has all
the __slots__ defined on it. This allows for multiple inheritance as
the immediate classes are simply support classes, with all the functions
and properties but none of the slots allocated.

Base.__new__ shows how one can swap out the support class with the true
data class (which inherits the support class).